temp: Rename .tsx -> .jsx & .ts -> .js for tests
This is a temporary commit since there are currently no webpack loaders that support tsx files in the test running. This commit should be removed once that is fixed upstream.
