+++ src/studio-home/tabs-section/libraries-v2-tab/index.jsx
